Allow customising the `jinja2.Environment`
Sometimes it might be desired to customise the Jinja Environment. For example, it might
be useful to enable the features in Jinja to strip/trim blocks to make it easier to control
whitespace rendering in templates.

Now it is possible to pass multiple parameters, which are fed into the `jinja2.Environment` as is.